Switch to deferred events to defend against thrown errors
This replaces all calls to EventEmitter#emit with a version that
emits the event in a setImmediate(). Previously, event handlers
that throw errors would result in them bubbling up the stack,
and effectively terminating the execution of the function
that emitted the event, which would break the internal machinery
of AMQPConnector and others. This is now fixed, and misbehaving
event listeners can no longer impact correct state transitions
such as reconnecting.

For #2